Fix -Wmaybe-uninitialized for gcc>=10
gcc>=10 now assumes that an enum can have any value, not just the defined
ones. This leads to many warnings after 'switch(enum)' constructs that have
no default case on purpose (to trigger -Wswitch warning).

To fix this, while preserving the -Wswitch behaviour, add a default case that
aborts and also force -Werror=switch-enum when using gcc (-Wswitch is silenced
when there is a default case). In the few cases where not all enum values must
be handled in a switch, cast the value to int to silence -Wswitch-enum.

+++ engine/tcl/component.c
@@ -387,7 +387,7 @@ comp_cmd(ClientData v, Tcl_Interp *interp, int objc, Tcl_Obj *const objv[])
       /* tasks */
       for(hash_first(comp_tasks(c), &i); i.current; hash_next(&i)) {
         for(hash_first(task_props(i.value), &j); j.current; hash_next(&j)) {
-          switch (prop_kind(j.value)) {
+          switch ((int)prop_kind(j.value)) {
             case PROP_VALIDATE: k = 1; break;
             case PROP_SIMPLE_CODEL: k = 2; break;
             case PROP_FSM_CODEL: k = 4; break;
@@ -403,7 +403,7 @@ comp_cmd(ClientData v, Tcl_Interp *interp, int objc, Tcl_Obj *const objv[])
       /* services */
       for(hash_first(comp_services(c), &i); i.current; hash_next(&i)) {
         for(hash_first(service_props(i.value), &j); j.current; hash_next(&j)) {
-          switch (prop_kind(j.value)) {
+          switch ((int)prop_kind(j.value)) {
             case PROP_VALIDATE: k = 1; break;
             case PROP_SIMPLE_CODEL: k = 2; break;
             case PROP_FSM_CODEL: k = 4; break;
